refactor(#155 PR 8/10): extract input/handlers.js from game.js
Pure code movement, zero behavior change. game.js: 2062 -> 1432 lines
(4076 at the start of the split).

- src/input/handlers.js (720 lines): the entire input layer — canvas
  raycast picking (getIntersect), wheel zoom, upgrade-indicator hover,
  keyboard state (keysPressed + blur reset), the mousedown/mousemove/
  mouseup drag/pan/connect/place handlers with all their state,
  tooltips, resize, keyboard shortcuts (Esc/H/R/T), toggleView and
  resetCamera. isIsometric moved with its writer (imported bindings
  are read-only). Bodies byte-verbatim, script-verified against HEAD.
- Listener registration is a module side effect; safe because no events
  can fire until the module graph finishes evaluating, and every
  game.js binding is dereferenced at event time only. animate() reads
  input state (keysPressed, isIsometric, ...) via live import bindings.
- tests/i18n-usage scans src/input, and restores src/sim which PR 7's
  test edit had silently dropped from key coverage.

Verified: lint 0, 87/87 tests, 30-module graph links, in-browser —
boot clean, a real mouse click through the moved handlers places a
service (toolbar tool -> canvas mousedown/mouseup), T/R keyboard
shortcuts and wheel zoom work, 40/40 requests processed end-to-end,
campaign level 2 prebuilds, zero console errors.
